requirements.txt
Flask==3.0.3
Flask_SQLAlchemy==3.1.1
SQLAlchemy>=2.0.43
Werkzeug==3.0.3
Pillow==10.4.0
python-dotenv==1.0.1
